23-june

+++ src/app/shopping-list/shopping-edit/shopping-edit.component.ts
@@ -1,4 +1,4 @@
-import { Component, OnInit, Output,EventEmitter, OnDestroy } from '@angular/core';
+import { Component, OnInit, Output, EventEmitter, OnDestroy } from '@angular/core';
 import { FormControl, FormGroup, Validators } from '@angular/forms';
 import { Subscription } from 'rxjs';
 
@@ -11,20 +11,20 @@ import { ShoppingListService } from '../shopping-list.service';
   styleUrls: ['./shopping-edit.component.css']
 })
 export class ShoppingEditComponent implements OnInit, OnDestroy {
-  shoppingListForm : FormGroup
-  forbiddenNames = ['test','testing']
-  subscription : Subscription
+  shoppingListForm: FormGroup
+  forbiddenNames = ['test', 'testing']
+  subscription: Subscription
   editMode = false
-  editedItemIndex : number
-  editedItem : Ingredient
+  editedItemIndex: number
+  editedItem: Ingredient
 
-  constructor(private shoppingListService : ShoppingListService) { }
+  constructor(private shoppingListService: ShoppingListService) { }
 
   ngOnInit(): void {
     //creating a new formgroup
     this.shoppingListForm = new FormGroup({
-      name: new FormControl('', [Validators.required,this.forbiddenName.bind(this)]),
-      amount: new FormControl('',[Validators.required,Validators.pattern('^[1-9]+[0-9]*$')])
+      name: new FormControl('', [Validators.required, this.forbiddenName.bind(this)]),
+      amount: new FormControl('', [Validators.required, Validators.pattern('^[1-9]+[0-9]*$')])
     })
 
     this.subscription = this.shoppingListService.startedEditing.subscribe(index => {
@@ -38,48 +38,48 @@ export class ShoppingEditComponent implements OnInit, OnDestroy {
     })
   }
 
-  ngOnDestroy(){
+  ngOnDestroy() {
     this.subscription.unsubscribe()
   }
 
-  onAdd(){
+  onAdd() {
     let ingredient: Ingredient = {
-        name : this.shoppingListForm.controls['name'].value,
-        amount : +this.shoppingListForm.controls['amount'].value
+      name: this.shoppingListForm.controls['name'].value,
+      amount: +this.shoppingListForm.controls['amount'].value
     }
     if (this.editMode != true)
       this.shoppingListService.addIngredient(ingredient)
-    this.shoppingListService.updateIngredient(ingredient,this.editedItemIndex)
+    this.shoppingListService.updateIngredient(ingredient, this.editedItemIndex)
 
     //console.log(this.shoppingListForm)
   }
 
-  onClear(){
+  onClear() {
     this.editMode = false
     this.editedItemIndex = null
     this.editedItem = null
     this.shoppingListForm.reset()
   }
 
-  onDelete(){
-
-    this.shoppingListService.deleteIngredient(this.editedItemIndex)
-
+  onDelete() {
+    if (this.editedItemIndex > -1) {
+      this.shoppingListService.deleteIngredient(this.editedItemIndex)
+    }
     this.editMode = false
-    this.editedItemIndex = null
+    this.editedItemIndex = -1
     this.editedItem = null
     this.shoppingListForm.reset()
   }
 
-  forbiddenName(control: FormControl): { [s:string] : boolean } {
-    if (this.forbiddenNames.indexOf(control.value) !== -1){
-      return { 'nameIsForbidden' : true}
+  forbiddenName(control: FormControl): { [s: string]: boolean } {
+    if (this.forbiddenNames.indexOf(control.value) !== -1) {
+      return { 'nameIsForbidden': true }
     }
     return null
 
   }
-  
-  
-  
+
+
+
 
 }
